feat(cloudsql/mysql): update to V2 sample
