+++ cloud-sql/mysql/mysql/app.standard.yaml
@@ -11,7 +11,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-runtime: nodejs10
+runtime: nodejs17
 
 # The following env variables may contain sensitive information that grants
 # anyone access to your database. Do not add this file to your source control.
