+++ cloud-sql/mysql/mysql/deployment.yaml
@@ -38,7 +38,7 @@ spec:
         env:
         - name: PORT
           value: "8080"
-        - name: DB_HOST
+        - name: INSTANCE_HOST
           value: "127.0.0.1"
         - name: DB_PORT
           value: "3306"
