+++ cloud-sql/mysql/mysql/views/index.pug
@@ -58,7 +58,7 @@ html(lang="en")
             }
           }
         };
-        xhr.open("POST", "/", true);
+        xhr.open("POST", "/votes", true);
         xhr.setRequestHeader("Content-Type", "application/x-www-form-urlencoded");
         xhr.send("team=" + team);
       }
